requirements.txt
flask==3.0.3
twilio==9.3.5
python-dotenv==1.0.1
openai==1.52.2
gunicorn==22.0.0
